Rename PingLogger to Logger
