+++ PingDavinci/PingDavinciTests/DaVinciErrorTests.swift
@@ -13,7 +13,7 @@ import Foundation
 import XCTest
 @testable import PingOrchestrate
 @testable import Storage
-@testable import PingLogger
+@testable import Logger
 @testable import PingOidc
 @testable import PingDavinci
