+++ PingDavinci/PingDavinciTests/integration tests/DaVinciIntegrationTests.swift
@@ -11,7 +11,7 @@
 
 import XCTest
 @testable import PingOrchestrate
-@testable import PingLogger
+@testable import Logger
 @testable import PingOidc
 @testable import Storage
 @testable import PingDavinci
