+++ PingLogger/Logger.xcodeproj/xcshareddata/IDETemplateMacros.plist
