+++ PingLogger/Logger/Logger.h
@@ -1,0 +1,21 @@
+//
+//  Logger.h
+//  Logger
+//
+//  Copyright (c) 2024 Ping Identity. All rights reserved.
+//
+//  This software may be modified and distributed under the terms
+//  of the MIT license. See the LICENSE file for details.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for Logger.
+FOUNDATION_EXPORT double LoggerVersionNumber;
+
+//! Project version string for Logger.
+FOUNDATION_EXPORT const unsigned char LoggerVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <Logger/PublicHeader.h>
+
+
