+++ PingLogger/Logger/PrivacyInfo.xcprivacy
