--- PingLogger/PingLogger/PingLogger.h
@@ -1,21 +1,0 @@
-//
-//  PingLogger.h
-//  PingLogger
-//
-//  Copyright (c) 2024 Ping Identity. All rights reserved.
-//
-//  This software may be modified and distributed under the terms
-//  of the MIT license. See the LICENSE file for details.
-//
-
-#import <Foundation/Foundation.h>
-
-//! Project version number for PingLogger.
-FOUNDATION_EXPORT double PingLoggerVersionNumber;
-
-//! Project version string for PingLogger.
-FOUNDATION_EXPORT const unsigned char PingLoggerVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <PingLogger/PublicHeader.h>
-
-
