+++ PingOrchestrate/PingOrchestrate/HttpClient.swift
@@ -10,7 +10,7 @@
 
 
 import Foundation
-import PingLogger
+import Logger
 
 /// `HttpClient` is responsible for handling HTTP requests and logging the details of those requests and responses.
 public class HttpClient {
